Minor fixes and improvements, also fixed some compiler warnings

+++ DV1573---UD1448/Renderer/Shader.cpp
@@ -244,7 +244,7 @@ void Shader::setMaterial(std::string materialName) {
 	setVec3("Diffuse_Color", mat->diffuse);
 	setVec3("Specular_Color", mat->specular);
 	
-	for (int i = 0; i < mat->textureID.size(); i++) {
+	for (size_t i = 0; i < mat->textureID.size(); i++) {
 		glActiveTexture(GL_TEXTURE0 + i);
 		glBindTexture(GL_TEXTURE_2D, mat->textureID.at(i));
